Move chunked uploads methods into a separate file
