+++ lib/boxr.rb
@@ -3,6 +3,7 @@ require 'httpclient'
 require 'hashie'
 require 'addressable/template'
 require 'jwt'
+require 'parallel'
 require 'securerandom'
 
 require 'boxr/version'
@@ -11,6 +12,7 @@ require 'boxr/client'
 require 'boxr/shared_items'
 require 'boxr/folders'
 require 'boxr/files'
+require 'boxr/chunked_uploads'
 require 'boxr/comments'
 require 'boxr/users'
 require 'boxr/groups'
